server/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
marshmallow==3.21.3
marshmallow-sqlalchemy==0.29.0
requests==2.32.3

